+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - UI for rendering dashboard with email tracking statistics
+- password protection for UI
 
 ## [0.2.3] - 2022-09-16
 ### Modified